requirements.txt
# To install: pip install -r requirements.txt
python-telegram-bot==21.3
google-generativeai==0.7.2
requests==2.32.3
aiohttp==3.11.11
